+++ src/main/java/dk/kb/kaltura/client/DsKalturaClient.java
@@ -11,12 +11,18 @@ import org.slf4j.LoggerFactory;
 import java.io.File;
 
 import com.kaltura.client.APIOkRequestsExecutor;
+import com.kaltura.client.Client;
 import com.kaltura.client.Configuration;
+import com.kaltura.client.enums.ESearchEntryFieldName;
+import com.kaltura.client.enums.ESearchItemType;
+import com.kaltura.client.enums.ESearchOperatorType;
 import com.kaltura.client.enums.MediaType;
 import com.kaltura.client.enums.SessionType;
+import com.kaltura.client.services.ESearchService;
 import com.kaltura.client.services.MediaService;
 import com.kaltura.client.services.MediaService.AddContentMediaBuilder;
 import com.kaltura.client.services.MediaService.AddMediaBuilder;
+import com.kaltura.client.services.MediaService.DeleteMediaBuilder;
 import com.kaltura.client.services.MediaService.ListMediaBuilder;
 import com.kaltura.client.services.UploadTokenService;
 import com.kaltura.client.services.UploadTokenService.AddUploadTokenBuilder;
@@ -32,7 +38,21 @@ import com.kaltura.client.types.UploadToken;
 import com.kaltura.client.types.UploadedFileTokenResource;
 import com.kaltura.client.types.MediaEntry;
 import com.kaltura.client.Client;
+import com.kaltura.client.types.*;
 import com.kaltura.client.utils.response.base.Response;
+
+import dk.kb.util.webservice.exception.InternalServiceException;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 
 /**
@@ -46,7 +66,10 @@ import com.kaltura.client.utils.response.base.Response;
  */
 public class DsKalturaClient {
 
-    static {        
+    // Kaltura-default: 30, maximum 500: https://developer.kaltura.com/api-docs/service/eSearch/action/searchEntry
+    public static final int BATCH_SIZE = 100;
+
+    static {
         // Kaltura library uses log4j2 and will remove this error message on start up: Log4j2 could not find a logging implementation
         System.setProperty("log4j2.loggerContextFactory", "org.apache.logging.log4j.simple.SimpleLoggerContextFactory");
     }
@@ -59,6 +82,7 @@ public class DsKalturaClient {
     private final String tokenId;
     private final long sessionKeepAliveSeconds;
     private long lastSessionStart=0;
+
     /**
      * Instantiate a session to Kaltura that can be used. The sessions can be reused between Kaltura calls without authenticating again.
      *
@@ -85,6 +109,24 @@ public class DsKalturaClient {
 
 
     /**
+     * <p>
+     * Delete a stream and all meta-data for the record in Kaltura.
+     * It can not be restored in Kaltura and must be uploaded again if deleted by mistake.
+     * </p>
+     *
+     * @param entryId The unique id in the Kaltura platform for the stream
+     * @return True if record was found and deleted. False if the record with the entryId could not be found in Kaltura.
+     * @throws IOException if Kaltura API called failed.
+     */
+    public boolean deleteStreamByEntryId(String entryId) throws IOException{
+         Client clientSession = getClientInstance();
+         DeleteMediaBuilder request = MediaService.delete(entryId);
+         Response<?> execute = APIOkRequestsExecutor.getExecutor().execute(request.build(clientSession)); // no object in response. Only status
+         return execute.isSuccess();
+    }
+
+
+    /**
      * Search Kaltura for a referenceId. The referenceId was given to Kaltura when uploading the record.<br>
      * We use filenames (file_id) as refereceIds. Example: b16bc5cb-1ea9-48d4-8e3c-2a94abae501b <br>
      * <br> 
@@ -101,7 +143,6 @@ public class DsKalturaClient {
 
         MediaEntryFilter filter = new MediaEntryFilter();
         filter.setReferenceIdEqual(referenceId);
-        //filter.idEqual("0_g9ys622b"); //Example to search for id
 
         FilterPager pager = new FilterPager();
         pager.setPageIndex(10);
@@ -109,9 +150,17 @@ public class DsKalturaClient {
         ListMediaBuilder request =  MediaService.list(filter);               
 
         //Getting this line correct was very hard. Little documentation and has to know which object to cast to.                
-        //For some documentation about the "Kaltura search" api see: https://developer.kaltura.com/api-docs/service/media/action/list                
+        //For some documentation about the "Kaltura search" api see: https://developer.kaltura.com/api-docs/service/media/action/list
         Response <ListResponse<MediaEntry>> response = (Response <ListResponse<MediaEntry>>) APIOkRequestsExecutor.getExecutor().execute(request.build(clientSession));
-        List<MediaEntry> mediaEntries = response.results.getObjects();           
+
+        //This is not normal situation. Normally Kaltura will return empty list: ({"objects":[],"totalCount":0,"objectType":"KalturaMediaListResponse"})
+        // When this happens something is wrong in kaltura and we dont know if there is results or not
+        if (response.results == null) {
+           log.error("Unexpected NULL response from Kaltura for referenceId:"+referenceId);
+            throw new InternalServiceException("Unexpected null response from Kaltura for referenceId:"+referenceId);
+        }
+
+        List<MediaEntry> mediaEntries = response.results.getObjects();
 
         int numberResults = mediaEntries.size();
 
@@ -128,6 +177,143 @@ public class DsKalturaClient {
     }
 
     /**
+     * Resolve Kaltura IDs for a list of referenceIDs.
+     * @param referenceIds a list of {@code referenceIDs}, typically UUIDs from stream filenames.
+     * @return a map from {@code referenceID} to {@code kalturaID}.
+     *         Unresolvable {@code referenceIDs} will not be present in the map.
+     * @throws IOException if the remote request failed.
+     */
+    public Map<String, String> getKulturaIds(List<String> referenceIds) throws IOException{
+        if (referenceIds.isEmpty()) {
+            log.info("getKulturaInternalIds(referenceIDs) called with empty list of IDs");
+            return Collections.emptyMap();
+        }
+
+        List<ESearchEntryBaseItem> items = referenceIds.stream()
+                .map(this::createReferenceIdItem)
+                .collect(Collectors.toList());
+        Response<ESearchEntryResponse> response = searchMulti(items);
+
+        // Collect result while checking for duplicates
+        final Map<String, String> pairs = new LinkedHashMap<>(referenceIds.size());
+        response.results.getObjects().stream()
+                .map(ESearchEntryResult::getObject)
+                .forEach(entry -> {
+                    String previousID;
+                    if ((previousID = pairs.put(entry.getReferenceId(), entry.getId())) != null) {
+                        log.warn("Warning: referenceID '{}' resolved to multiple kalturaIDs [\"{}\", \"{}\"]",
+                                entry.getReferenceId(), previousID, entry.getId());
+                    }});
+        return pairs;
+    }
+
+    /**
+     * Resolve referenceIDs for a list of Kaltura IDs.
+     * @param kalturaIDs a list of {@code kalturaIDs}.
+     * @return a map from {@code kalturaID} to {@code referenceID}.
+     *         Unresolvable {@code kalturaIDs} will not be present in the map.
+     * @throws IOException if the remote request failed.
+     */
+    public Map<String, String> getReferenceIds(List<String> kalturaIDs) throws IOException{
+        if (kalturaIDs.isEmpty()) {
+            log.info("getReferenceIds(kalturaIDs) called with empty list of IDs");
+            return Collections.emptyMap();
+        }
+
+        List<ESearchEntryBaseItem> items = kalturaIDs.stream()
+                .map(this::createKalturaIdItem)
+                .collect(Collectors.toList());
+        Response<ESearchEntryResponse> response = searchMulti(items);
+
+        return response.results.getObjects().stream()
+                        .map(ESearchEntryResult::getObject)
+                .collect(Collectors.toMap(BaseEntry::getId, BaseEntry::getReferenceId));
+    }
+
+    /**
+     * Simple free form term search in Kaltura.
+     * @param term a search term, such as {@code dr} or {@code tv avisen}.
+     * @return a list of Kaltura IDs for matching records, empty if no hits. Max result size is {@link #BATCH_SIZE}.
+     * @throws IOException if the remote request failed.
+     */
+    public List<String> searchTerm(String term) throws IOException{
+        // Adapted from Java samples at https://developer.kaltura.com
+        // https://developer.kaltura.com/console/service/eSearch/action/searchEntry?query=search
+        // https://developer.kaltura.com/api-docs/Search--Discover-and-Personalize/esearch.html
+        // TODO: This retrieves the full item representation. How to reduce to only [id, referenceId] fields?
+
+        ESearchUnifiedItem item = new ESearchUnifiedItem();
+        item.setItemType(ESearchItemType.EXACT_MATCH);
+        item.searchTerm(term);
+
+        return searchMulti(List.of(item)).results.getObjects().stream()
+                .map(ESearchEntryResult::getObject)
+                .map(BaseEntry::getId)
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * Generic multi search for a list of {@link ESearchEntryBaseItem items},
+     * returning at most {@link #BATCH_SIZE} results.
+     * @param items at least 1 search item.
+     * @return the response from a Kaltura search for the given items.
+     * @throws IOException if the remote request failed.
+     */
+    @SuppressWarnings("unchecked")
+    private Response<ESearchEntryResponse> searchMulti(List<ESearchEntryBaseItem> items) throws IOException{
+        // Adapted from Java samples at https://developer.kaltura.com
+        // https://developer.kaltura.com/console/service/eSearch/action/searchEntry?query=search
+        // https://developer.kaltura.com/api-docs/Search--Discover-and-Personalize/esearch.html
+        // TODO: This retrieves the full item representation. How to reduce to only [id, referenceId] fields?
+
+        if (items.size() > BATCH_SIZE) {
+            // TODO: Change this to multiple batch requests
+            throw new IllegalArgumentException(
+                    "Request for " + items.size() + " items exceeds current limit of " + BATCH_SIZE);
+        }
+
+        // Setup request
+        ESearchEntryParams searchParams = new ESearchEntryParams();
+        ESearchEntryOperator operator = new ESearchEntryOperator();
+        operator.setOperator(ESearchOperatorType.OR_OP);
+        searchParams.setSearchOperator(operator);
+        operator.setSearchItems(items);
+        FilterPager pager = new FilterPager();
+        pager.setPageSize(BATCH_SIZE);
+
+        // Issue search
+        ESearchService.SearchEntryESearchBuilder requestBuilder = ESearchService.searchEntry(searchParams, pager);
+        return (Response<ESearchEntryResponse>)
+                APIOkRequestsExecutor.getExecutor().execute(requestBuilder.build(getClientInstance()));
+    }
+
+    /**
+     * Build a search item aka search clause for the given {@code referenceId}.
+     * @param referenceID typically the UUID for a stream filename.
+     * @return a search item ready for search or for building more complex search requests.
+     */
+    private ESearchEntryItem createReferenceIdItem(String referenceID) {
+        ESearchEntryItem item = new ESearchEntryItem();
+        item.setFieldName(ESearchEntryFieldName.REFERENCE_ID);
+        item.searchTerm(referenceID);
+        item.setItemType(ESearchItemType.EXACT_MATCH);
+        return item;
+    }
+
+    /**
+     * Build a search item aka search clause for the given {@code ID}.
+     * @param kalturaID typically the UUID for a stream filename.
+     * @return a search item ready for search or for building more complex search requests.
+     */
+    private ESearchEntryItem createKalturaIdItem(String kalturaID) {
+        ESearchEntryItem item = new ESearchEntryItem();
+        item.setFieldName(ESearchEntryFieldName.ID);
+        item.searchTerm(kalturaID);
+        item.setItemType(ESearchItemType.EXACT_MATCH);
+        return item;
+    }
+
+    /**
      * Upload a video or audio file to Kaltura. 
      * The upload require 4 API calls to Kaltura
      * <p><ul>
@@ -141,9 +327,9 @@ public class DsKalturaClient {
      * seem possible to later see the file in the kaltura administration gui. This error has only happened because I forced it. 
      * 
      * @param filePath File path to the media file to upload. 
-     * @param referenceId Use our internal ID's there. This referenceId can be used to find the record at Kaltura and also map to internal KalturaId.
+     * @param referenceId. Use our internal ID's there. This referenceId can be used to find the record at Kaltura and also map to internal KalturaId.
      * @param mediaType enum type. MediaType.AUDIO or MediaType.VIDEO
-     * @param title Name/titel for the resource in Kaltura
+     * @param name Name/titel for the resource in Kaltura
      * @param description , optional description 
      * @param tag Optional tag. Uploads from the DS should always use tag 'DS-KALTURA'.  There is no backup for this tag in Kaltura and all uploads can be deleted easy.
      *
@@ -250,7 +436,9 @@ public class DsKalturaClient {
     private synchronized Client getClientInstance() throws IOException{
         try {
 
-            if (System.currentTimeMillis()-lastSessionStart >= sessionKeepAliveSeconds/1000) {
+            if (System.currentTimeMillis()-lastSessionStart >= sessionKeepAliveSeconds*1000) {
+                //Create the client
+                //KalturaConfiguration config = new KalturaConfiguration();
                 Configuration config = new Configuration();
                 config.setEndpoint(kalturaUrl);
                 Client client = new Client(config);