+++ CHANGELOG.md
@@ -5,9 +5,31 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
 
+### Added
+
+
+### Changed
+- Bump kb-util to v1.5.11
+
+
+## [1.2.3] (https://github.com/kb-dk/ds-kaltura/releases/tag/ds-kaltura-1.2.3) - 2024-07-18
+- Handle rare error from Kaltura with empty response.
+
+## [1.2.2] (https://github.com/kb-dk/ds-kaltura/releases/tag/ds-kaltura-1.2.2) - 2024-07-17
+- New manuel delete job that takes an input file of kaltura entryId's and deletes each stream+metadata in Kaltura for the entries.
+
+## [1.2.1] (https://github.com/kb-dk/ds-kaltura/releases/tag/ds-kaltura-1.2.1) - 2024-06-27
+### Added
+- New method that deletes a stream and all metadata for the stream in Kaltura by the internal Kaltura entryId as key.
+
+### Changed
+- Updated kb-util version
+
+
+## [1.2.0] (https://github.com/kb-dk/ds-kaltura/releases/tag/ds-kaltura-1.2.0) - 2024-05-28 - 2024-06-21
 - Remove the root `.config` from the configuration
+- More methods.
 
 ## [1.0.0] - YYYY-MM-DD
 ### Added